+++ .metadata
@@ -4,7 +4,27 @@
 # This file should be version controlled and should not be manually edited.
 
 version:
-  revision: 097d3313d8e2c7f901932d63e537c1acefb87800
-  channel: stable
+  revision: "2524052335ec76bb03e04ede244b071f1b86d190"
+  channel: "stable"
 
 project_type: app
+
+# Tracks metadata for the flutter migrate command
+migration:
+  platforms:
+    - platform: root
+      create_revision: 2524052335ec76bb03e04ede244b071f1b86d190
+      base_revision: 2524052335ec76bb03e04ede244b071f1b86d190
+    - platform: ios
+      create_revision: 2524052335ec76bb03e04ede244b071f1b86d190
+      base_revision: 2524052335ec76bb03e04ede244b071f1b86d190
+
+  # User provided section
+
+  # List of Local paths (relative to this file) that should be
+  # ignored by the migrate tool.
+  #
+  # Files that are not part of the templates will be ignored by default.
+  unmanaged_files:
+    - 'lib/main.dart'
+    - 'ios/Runner.xcodeproj/project.pbxproj'
